[Feat] consumed_foods 테이블 엔티티 생성

+++ src/main/java/com/daall/howtoeat/user/UserTargets.java
@@ -22,7 +22,7 @@ public class UserTargets extends Timestamped {
     private User user;
 
     @Column(nullable = false)
-    private Integer kcal;
+    private Double kcal;
 
     @Column(nullable = false)
     private Double carbo;
